Update server command in tasks.json, add build script for PyInstaller, and include PyInstaller in requirements
