+++ .vscode/tasks.json
@@ -4,7 +4,7 @@
         {
             "label": "Start Python HTTP Server",
             "type": "shell",
-            "command": "python3 server.py",
+            "command": "python3 serve.py",
             "group": {
                 "kind": "build",
                 "isDefault": true
