+++ requirements.txt
@@ -3,3 +3,4 @@ beautifulsoup4
 waitress
 regex
 requests
+pyinstaller
